refactored api to smaller modules and updated tests with correct stubbing
