+++ catalog/main_api.py
@@ -1,0 +1,9 @@
+from catalog.api.api import create_app
+
+def main():
+    app = create_app()
+    app.run(host="127.0.0.1", port=5000)
+
+
+if __name__ == "__main__":
+    main()
